requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
python-multipart==0.0.20
streamlit==1.44.1
pymongo==4.12.0
langchain==0.3.24
langchain-community==0.3.22
langchain-chroma==0.2.3
langchain-google-genai==2.1.3
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.8
chromadb==0.6.3
huggingface_hub==0.30.2
sentence-transformers==4.1.0
transformers==4.51.3
pypdf==5.4.0
docx2txt==0.9
unstructured==0.17.2
SpeechRecognition==3.10.4
python-dotenv==1.1.0
requests==2.32.3
pydantic==2.11.3